Corrected wrong tax calculation on shipping
The calculation of the tax percent was based on the total shipping amount (including the tax), but this should be based on what percentage it is of the shipping price excluding tax. Otherwise this will throw an error. In the epay window.

For example:

if the shipping price including tax is 100 DKK, and the tax is 25% that means the price excluding tax is 80 DKK.
So with previous calculation it would be 20/100*100 that would be 20%
With updated calculation it would be 20/80*100 which is 25% and the correct amount.